feat: make lazy loading default on checkouts
