+++ support-frontend/webpack.entryPoints.js
@@ -1,7 +1,6 @@
 module.exports = {
 	common: {
 		'[countryGroupId]/router': 'pages/[countryGroupId]/router.tsx',
-		'[countryGroupId]/lazyRouter': 'pages/[countryGroupId]/lazyRouter.tsx',
 		'[countryGroupId]/events/router':
 			'pages/[countryGroupId]/events/router.tsx',
 		favicons: 'images/favicons.ts',
